@@ -1,17 +1,21 @@
 {
-	"name": "sbst",
+	"name": "substrate-next",
 	"version": "0.0.1",
+	"private": true,
 	"scripts": {
-		"dev": "vite dev --host",
+		"dev": "vite dev",
 		"build": "vite build",
-		"preview": "vite preview --host",
-		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json"
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"lint": "eslint ."
 	},
 	"devDependencies": {
-		"@sveltejs/kit": "^1.27.3",
-		"@types/seedrandom": "^3.0.7",
-		"svelte": "^4.2.2",
-		"svelte-check": "^3.5.2",
+		"@sveltejs/kit": "^1.27.6",
+		"@typescript-eslint/eslint-plugin": "^6.11.0",
+		"@typescript-eslint/parser": "^6.11.0",
+		"eslint": "^8.53.0",
+		"eslint-plugin-svelte": "^2.35.0",
+		"svelte": "5.0.0-next.2",
+		"svelte-check": "^3.6.0",
 		"tslib": "^2.6.2",
 		"typescript": "^5.2.2",
 		"vite": "^4.5.0"
@@ -19,7 +23,7 @@
 	"type": "module",
 	"dependencies": {
 		"@sveltejs/adapter-static": "^2.0.3",
-		"seedrandom": "^3.0.5",
-		"svelte-persisted-store": "^0.7.0"
+		"@types/seedrandom": "^3.0.8",
+		"seedrandom": "^3.0.5"
 	}
 }
